refactor(voip): split VoiceNoteScreen view model and recording logic
